Offset Correction
It seems some of the measurements have a bad starting displacement value. Now you can toggle on and off aligning the first non-zero force at 0mm.
Technically that is not completely accurate as it looks for the maximum displacement with a force of 0, regardless of up or down stroke. We can expect symmetry though, so this should be acceptable.
Also some other minor cleaning tweaks